requirements.txt
pillow
zipfile
